Refactor for clean usage of httputil
