+++ gateway/event_handler_jsvm_test.go
@@ -1,8 +1,9 @@
 package gateway
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestJSVMEventHandler_Init(t *testing.T) {
