+++ gateway/event_system_test.go
@@ -2,10 +2,11 @@ package gateway
 
 import (
 	"bytes"
-	"github.com/stretchr/testify/assert"
 	"strings"
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
 
 	"github.com/sirupsen/logrus"
