+++ internal/service/core/context.go
@@ -1,4 +1,4 @@
-package httputil
+package core
 
 import (
 	"context"
