+++ internal/service/core/context_test.go
@@ -1,13 +1,12 @@
-package httputil_test
+package core_test
 
 import (
 	"context"
 	"net/http"
 	"testing"
 
-	"github.com/TykTechnologies/tyk/internal/httputil"
-
 	"github.com/TykTechnologies/tyk/internal/model"
+	"github.com/TykTechnologies/tyk/internal/service/core"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -24,10 +23,10 @@ func TestUpstreamAuth(t *testing.T) {
 		mockAuthProvider := &model.MockUpstreamAuthProvider{}
 		req := createReq(t)
 
-		httputil.SetUpstreamAuth(req, mockAuthProvider)
+		core.SetUpstreamAuth(req, mockAuthProvider)
 
 		// Retrieve the auth provider from the request's context to verify it was set
-		retrievedAuth := httputil.GetUpstreamAuth(req)
+		retrievedAuth := core.GetUpstreamAuth(req)
 		assert.NotNil(t, retrievedAuth)
 		assert.Equal(t, mockAuthProvider, retrievedAuth)
 	})
@@ -35,7 +34,7 @@ func TestUpstreamAuth(t *testing.T) {
 	t.Run("no auth provider", func(t *testing.T) {
 		req := createReq(t)
 
-		retrievedAuth := httputil.GetUpstreamAuth(req)
+		retrievedAuth := core.GetUpstreamAuth(req)
 		assert.Nil(t, retrievedAuth)
 	})
 
@@ -43,10 +42,10 @@ func TestUpstreamAuth(t *testing.T) {
 		req := createReq(t)
 
 		// Set a context with a value that is not of type proxy.UpstreamAuthProvider
-		ctx := context.WithValue(req.Context(), httputil.ContextKey("upstream-auth"), "invalid-type")
-		httputil.SetContext(req, ctx)
+		ctx := context.WithValue(req.Context(), core.ContextKey("upstream-auth"), "invalid-type")
+		core.SetContext(req, ctx)
 
-		retrievedAuth := httputil.GetUpstreamAuth(req)
+		retrievedAuth := core.GetUpstreamAuth(req)
 		assert.Nil(t, retrievedAuth)
 	})
 }
@@ -56,30 +55,30 @@ func TestSetContext(t *testing.T) {
 		req := createReq(t)
 
 		// Create a new context with a key-value pair
-		ctx := context.WithValue(context.Background(), httputil.ContextKey("key"), "value")
+		ctx := context.WithValue(context.Background(), core.ContextKey("key"), "value")
 
 		// Call SetContext to update the request's context
-		httputil.SetContext(req, ctx)
+		core.SetContext(req, ctx)
 
 		// Verify that the request's context has been updated
-		retrievedValue := req.Context().Value(httputil.ContextKey("key"))
+		retrievedValue := req.Context().Value(core.ContextKey("key"))
 		assert.Equal(t, "value", retrievedValue)
 	})
 
 	t.Run("override key", func(t *testing.T) {
 
 		req := createReq(t)
-		existingCtx := context.WithValue(context.Background(), httputil.ContextKey("existingKey"), "existingValue")
+		existingCtx := context.WithValue(context.Background(), core.ContextKey("existingKey"), "existingValue")
 		req = req.WithContext(existingCtx)
 
 		// Create a new context to override the existing context
-		newCtx := context.WithValue(context.Background(), httputil.ContextKey("newKey"), "newValue")
+		newCtx := context.WithValue(context.Background(), core.ContextKey("newKey"), "newValue")
 
 		// Call SetContext to update the request's context with the new context
-		httputil.SetContext(req, newCtx)
+		core.SetContext(req, newCtx)
 
-		assert.Nil(t, req.Context().Value(httputil.ContextKey("existingKey")))
-		assert.Equal(t, "newValue", req.Context().Value(httputil.ContextKey("newKey")))
+		assert.Nil(t, req.Context().Value(core.ContextKey("existingKey")))
+		assert.Equal(t, "newValue", req.Context().Value(core.ContextKey("newKey")))
 	})
 
 	t.Run("empty context", func(t *testing.T) {
@@ -87,7 +86,7 @@ func TestSetContext(t *testing.T) {
 
 		emptyCtx := context.Background()
 
-		httputil.SetContext(req, emptyCtx)
+		core.SetContext(req, emptyCtx)
 
 		assert.Equal(t, emptyCtx, req.Context())
 	})
